Node.js上での実行を想定し、モジュール化してjs実装
