+++ index.html
@@ -53,6 +53,6 @@
             </div>
         </div>
     </div>
-    <script src="js/main.js"></script>
+    <script type="module" src="js/main.js"></script>
 </body>
 </html>
